+++ src/Domain/Model/Visio/VisioSettings.php
@@ -29,6 +29,9 @@ class VisioSettings
                 new VisioSettingsTranslation(
                     $this,
                     $locale,
+                    null,
+                    null,
+                    null,
                     null
                 )
             );
@@ -74,9 +77,72 @@ class VisioSettings
         return null !== $translation->getHeader();
     }
 
+    public function hasEndSound(string $locale): bool
+    {
+        /** @var VisioSettingsTranslation $translation */
+        $translation = $this->translations->get($locale);
+
+        if (null === $translation) {
+            return false;
+        }
+
+        return null !== $translation->getEndSound();
+    }
+
+    public function getEndSound(string $locale): ?string
+    {
+        /** @var VisioSettingsTranslation $translation */
+        $translation = $this->translations->get($locale);
+
+        if (null === $translation) {
+            return null;
+        }
+
+        return $translation->getEndSound();
+    }
+
+    public function hasEndImage(string $locale): bool
+    {
+        /** @var VisioSettingsTranslation $translation */
+        $translation = $this->translations->get($locale);
+
+        if (null === $translation) {
+            return false;
+        }
+
+        return null !== $translation->getEndImage();
+    }
+
+    public function getEndImage(string $locale): ?string
+    {
+        /** @var VisioSettingsTranslation $translation */
+        $translation = $this->translations->get($locale);
+
+        if (null === $translation) {
+            return null;
+        }
+
+        return $translation->getEndImage();
+    }
+
+    public function getEndMessage($locale): ?string
+    {
+        /** @var VisioSettingsTranslation $translation */
+        $translation = $this->translations->get($locale);
+
+        if (null === $translation) {
+            return null;
+        }
+
+        return $translation->getEndMessage();
+    }
+
     public function updateTranslation(
         string $locale,
-        ?string $header
+        ?string $header,
+        ?string $endSound,
+        ?string $endImage,
+        ?string $endMessage
     ): void {
         /** @var VisioSettingsTranslation $translation */
         $translation = $this->translations->get($locale);
@@ -84,12 +150,24 @@ class VisioSettings
         if (null === $translation) {
             $this->translations->set(
                 $locale,
-                new VisioSettingsTranslation($this, $locale, $header)
+                new VisioSettingsTranslation(
+                    $this,
+                    $locale,
+                    $header,
+                    $endSound,
+                    $endImage,
+                    $endMessage
+                )
             );
 
             return;
         }
 
-        $translation->update($header);
+        $translation->update(
+            $header,
+            $endSound,
+            $endImage,
+            $endMessage
+        );
     }
 }