+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20200518091421.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Add endSound, endImage and endMessage on VisioSettings.
+ */
+final class Version20200518091421 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE visio_settings_translation ADD end_sound VARCHAR(255) DEFAULT NULL, ADD end_image VARCHAR(255) DEFAULT NULL, ADD end_message LONGTEXT DEFAULT NULL');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE visio_settings_translation DROP end_sound, DROP end_image, DROP end_message');
+    }
+}