+++ src/Ui/Bundle/EventBundle/Form/Type/Participant/SortParticipantsType.php
@@ -1,0 +1,44 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Form\Type\Participant;
+
+use Proximum\Vimeet\Application\Command\Sheet\SortParticipants;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class SortParticipantsType extends AbstractType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        /** @var SortParticipants $sortParticipants */
+        $sortParticipants = $options['sortParticipants'];
+
+        foreach ($sortParticipants->getParticipantsRank() as $id => $rank) {
+            $builder->add($id, IntegerType::class);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'data_class' => SortParticipants::class,
+        ]);
+        $resolver->setRequired(['sortParticipants']);
+    }
+}